requirements.txt




fastapi
uvicorn

Flask==2.1.2

rudalle==1.1.1
dalle-pytorch==0.14.3
tqdm
deepspeed==0.4.2
triton==0.4.2